+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `postpic.Field` method `mean` has now an interface matching `ndarray.mean`. This means that, if the `axis` argument is not given, it averages across all axes instead the last axis.
 * `postpic.Field.map_coordinates` applies now the Jacobian determinant of the transformation, in order to preserve the definite integral.
 In your code you will need to turn calls to `Field.transform` into calls to `Field.map_coordinates` and set the keyword argument `preserve_integral=False` to get the old behaviour.
+* `postpic.MultiSpecies.createField` expects the histogram arguments (`bins`, `shape`) directly now. Before this change they were expected as a separate dictionary `optargsh`.
 * The functions `MultiSpecies.compress`, `MultiSpecies.filter`, `MultiSpecies.uncompress` and `ParticleHistory.skip` return a new object now. Before this release, they modified the current object. Assuming  `ms` is a `MultiSpecies` object, the corresponding adjustemens read:<br>
 old: `ms.filter('gamma > 2')`<br>
 new: `ms = ms.filter('gamma > 2')`